Con confirmación

+++ src/app/features/users/components/users-list/users-list.component.ts
@@ -1,7 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { UsersService } from '../../services/users.service';
 import { User } from '../../models/user.model';
-
+import { MatDialog } from '@angular/material/dialog';
+import { MiDialogoComponent } from 'src/app/shared/mi-dialogo/mi-dialogo.component';
 @Component({
   selector: 'app-users-list',
   templateUrl: './users-list.component.html',
@@ -9,7 +10,7 @@ import { User } from '../../models/user.model';
 })
 export class UsersListComponent implements OnInit {
   usersList: User[] = [];
-  constructor(private usersService: UsersService) {}
+  constructor(private usersService: UsersService, private dialogo:MatDialog ) {}
   ngOnInit(): void {
     this.getLista();
   }
@@ -26,9 +27,28 @@ export class UsersListComponent implements OnInit {
    * @param id 
    */
 delete(id:string):void{
-  this.usersService.delete(id).subscribe((usuarios)=>{
-    this.usersList=usuarios
-  })
+  
+  const dialogRef = this.dialogo.open(MiDialogoComponent, {
+    panelClass:'mi-dialogo-personalizado',
+    data: {
+      titulo: `Eliminar el usuario ${id}`,
+      mensaje:'¿Estás seguro?',
+      submensaje:'Esto es una prueba'
+    }
+  });
+  dialogRef.afterClosed().subscribe(result => {
+    if (result) {
+      
+      this.usersService.delete(id).subscribe((usuarios)=>{
+        this.usersList=usuarios
+      })
+    }
+  });
+
+
+
+
+  
 }
 
 }
